env_state: Use chain of meta nodes for designation
Prior to this change, designation only occurred in the current
environment. If we want to keep the default working environment
separate from the lang environment, EnvState must be capable of
designating over multiple environments.

This change currently leaves designation chain policy entirely up to
clients. In the future, some sort of interface that at least ensures
the validity of inserted nodes should be used. Holding off on that
for now since that interface may end up being built in amlang.