interactive_stream: Indicate when line doesn't complete expression
This change also compresses multi-line expressions into a single
editor history entry.

One missing functionality is depth changes for quotes. Since tokenize
doesn't factor in depth of quotes, top-level quotes will not lead to
an indication that the next line is part of an existing
expression. This does not manifest as an issue for depths below
top-level.