env_manager: Move policy functionality into EnvPolicy
EnvPolicy serves to allow for amlang to be used as a library without
constraining the kinds of deployments clients can use (e.g. monolithic
vs microservice). This is important because the Environment hierarchy
represents the major component of shared state amongst Agents.

This change also changes EnvManager bootstrapping to be in-place
rather than creating multiple EnvManager instances to get to the final
bootstrapped version.