sexp_conversion: break_by_types: Avoid extra TryFrom calls
Prior to this change, break_by_types! would first test a conversion
using <&T>::try_from(&sexp) and then reconvert with
<T>::try_from(sexp) on success. This was done because failing a
<T>::try_from(sexp) conversion previously irrevocably consumed the
sexp.

Now that the sexp is returned within the TryFrom Err, we can avoid the
"test" TryFrom conversion. Not only is this more efficient, but it
means the only constraint on $type is now TryFrom<$type> for Sexp,
whereas before we also required TryFrom<&$type> for &Sexp.