sexp: SexpIntoIter from Sexp without moving to the heap
SexpIntoIter presented an awkward situation, since the car/cdr of Cons
will always be HeapSexp, but the top-level s-exp may be a Sexp rather
than HeapSexp.

This CL resolves the awkwardness without requiring clients to manually
convert to HeapSexp. Indeed, in normal cases there will be no
conversion to a HeapSexp. The only cases in which the top-level Sexp
will be converted to a HeapSexp are:
  * Converting a Primitive s-exp to a SexpIntoIter & calling next().
  * Converting a s-exp to a SexpIntoIter & immediately calling
    consume().