agent_state: Concretize into Sexps
Concretizing into Nodes implied that relevant execution values will
always be in an Environment. Returning Sexps allows for execution
values to be stored temporarily in a Continuation or in some other
form of transient state.

This change does not tackle integrating value reference lifetime
constraints based on, say, Continuation frame lifetimes. Rather, we
simply clone values during concretization. While not ideal, this
doesn't interfere with the ability to efficiently compile down
Procedures, so it's not a big deal for now.