cargo: Allow downstream users to accept minor & patch changes
Essentially, we want to provide downstream users with the ability to
use any version of dependencies that doesn't break us, without
completely losing the ability for amlang development to fully specify
the versioning for tests & examples.

Committing the Cargo.lock also enables downstream users to constrain
their builds to tested & trusted (as much as dependencies can be
trusted) versioning. On the other hand, without going out of their way
to use this versioning, downstream users will not be constrained.