environment: mem_backend: Remove log::trace calls
We're seeing stack overflows in certain cases, where logging is
setup by a different framework which isn't ignoring trace calls. In
any case, this is probably too low-level to be of value even as trace
calls; if we really want to step through these functions, rust-gdb is
probably the correct approach.