stream: strategy: Templatize for static dispatch
Using Stream as a boundary for dynamic dispatch should take care of
a dynamic Strategy's Input half. Something similar can be done with
the Transform half when needed. On the other hand, normal static usage
shouldn't have to pay the cost for extra dynamic dispatch for now. If
we really determine we want to reify those base details in the future,
we can always do so.