env: Have MetaEnv store envs separately
Having Box<EnvObject> be a kind of Primitive brought problems with
Send + Sync (among other issues) and had limited benefit bc the
concept of a meta env neatly bounded the scope and envs can simply be
reified into Sexps if an Agent needs to interact with that.

We'll have to deal with concurrency-related issues separately for
MetaEnv.envs, but mutating the structure should be infrequent enough
that even locking would work (or may be completely unnecessary
depending on operational setup).