env: Add designation interface
Prior to this change, an amlang_designator SymNodeTable would be
populated at env creation, and then mutated by Agents at runtime.

This approach polluted the env unnecessarily (we rarely want that
reified), and placing the mutation at the Agent-level precludes sane
concurrency behavior (which should be encapsulated at the env level).

This change doesn't change the Agent-level semantics, but provides the
support at the env-level.