sexp: list!: Improved ()/nest support
Deep nesting still doesn't work as desired; we need to add commas so
the expression will be parsed as a tuple type rather than duplicate
parens. But then the macro will greedily munch as an expression,
forcing us to implement a bunch of From<> for Sexp, for tuple types
composed of unit types.

For now, manually making multiple calls to list! works around that
issue.