version: Initial commit
Given that we need versioning of rust API, env format, and different
interpreters & contexts, this change abstracts out version semantics
in a way that can both be tersely represented in a string and
explicitly laid out in a way that can be reified.

This change doesn't actually sort out migration/interop, it just
provides a repr.