primitive: From<str-like> for sexp-like
Prior to this change, we had a ToLangString trait, and users had to
specify whether they wanted a str-like element to be string or symbol.

This change moves semantics so that str-like elements can by directly
wrapped into a LangString; users only need to go out of their way if
they want a Symbol.