env_manager: Env 0.0.3 saving
Having a top-level nodes and triples sexp means that parsing needs to
allocate memory for the whole structure and produce one large sexp.
Let's use section headers instead so that we get many small top-level
sexps.

Ultimately a lot of this code should be serde of a file backend to the
env interface, but one thing at a time.

AMLANG: (migrate env ser 0.0.3)